requirements.txt
agentzero==0.4.4
bcrypt==3.1.7
chemist==1.6.3
click==7.0
coloredlogs==10.0
elasticsearch==7.5.1
flask-cors==3.0.8
flask-restplus==0.13.0
flask-session==0.3.1
flask==1.1.1
gevent==1.4.0
psycopg2-binary==2.8.4
pyzmq==18.0.1
redis==3.3.11
requests==2.22.0
uwsgi==2.0.18
werkzeug==0.16.1
